moving activemq to a new dedicated maven module
